Move test since there was a typo
