+++ tests/src/Rector/Deprecation/ConstantToClassConstantRector/config/configured_rule.php
