+++ tests/src/Rector/Deprecation/ConstantToClassConstantRector/fixture/fixture.php.inc
